Move bundled-skills install/status logic into @invoker/shell
packages/app/src/ipc/gui-mutation-handlers.ts imported
installBundledSkills/resolveBundledSkillsStatus from @invoker/cli/bundled-skills,
which violates the layer-4-deps dependency-cruiser rule (app may not depend on
cli). Move the shared logic (and the commandExists probe it depends on) into
the layer-0 @invoker/shell package so both cli and app import from an allowed
location; packages/cli/src/bundled-skills.ts now re-exports from there.
