+++ packages/shell/src/command-exists.ts
@@ -1,0 +1,5 @@
+import { spawnSync } from 'node:child_process';
+
+export function commandExists(command: string): boolean {
+  return spawnSync('sh', ['-c', `command -v ${command} >/dev/null 2>&1`], { stdio: 'ignore' }).status === 0;
+}
