+++ packages/shell/src/index.ts
@@ -1,1 +1,3 @@
 // @invoker/shell - Shell command execution and process management
+
+export * from './command-exists.ts';
